requirements-pdf.txt
# PDF Processing Dependencies
pdfplumber>=0.10.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
opencv-python>=4.8.0
PyMuPDF>=1.23.0  # For extracting embedded images from PDFs

